+++ x/programs/examples/util.go
@@ -5,8 +5,11 @@ package examples
 
 import (
 	"context"
+	"os"
 
+	"github.com/ava-labs/avalanchego/database/memdb"
 	"github.com/ava-labs/hypersdk/crypto/ed25519"
+	"github.com/ava-labs/hypersdk/state"
 	"github.com/ava-labs/hypersdk/x/programs/runtime"
 )
 
@@ -32,4 +35,43 @@ func newKey() (ed25519.PrivateKey, ed25519.PublicKey, error) {
 	}
 
 	return priv, priv.PublicKey(), nil
+}
+
+var (
+	_ state.Mutable   = &testDB{}
+	_ state.Immutable = &testDB{}
+)
+
+type testDB struct {
+	db *memdb.Database
+}
+
+func newTestDB() *testDB {
+	return &testDB{
+		db: memdb.New(),
+	}
+}
+
+func (c *testDB) GetValue(_ context.Context, key []byte) ([]byte, error) {
+	return c.db.Get(key)
+}
+
+func (c *testDB) Insert(_ context.Context, key []byte, value []byte) error {
+	return c.db.Put(key, value)
+}
+
+func (c *testDB) Put(key []byte, value []byte) error {
+	return c.db.Put(key, value)
+}
+
+func (c *testDB) Remove(_ context.Context, key []byte) error {
+	return c.db.Delete(key)
+}
+
+func GetProgramBytes(filePath string) ([]byte, error) {
+	return os.ReadFile(filePath)
+}
+
+func GetGuestFnName(name string) string {
+	return name + "_guest"
 }
