+++ x/programs/runtime/errors.go
@@ -14,4 +14,6 @@ var (
 	ErrInvalidParamCount            = errors.New("invalid parameter count")
 	ErrInvalidParamType             = errors.New("invalid parameter type")
 	ErrInsufficientUnits            = errors.New("insufficient units")
+	ErrRuntimeStoreSet              = errors.New("runtime store has already been set")
+	ErrNegativeValue                = errors.New("negative value")
 )
