--- x/programs/utils/utils.go
@@ -1,51 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package utils
-
-import (
-	"context"
-	"os"
-
-	"github.com/ava-labs/avalanchego/database/memdb"
-	"github.com/ava-labs/hypersdk/state"
-)
-
-var (
-	_ state.Mutable   = &testDB{}
-	_ state.Immutable = &testDB{}
-)
-
-type testDB struct {
-	db *memdb.Database
-}
-
-func NewTestDB() *testDB {
-	return &testDB{
-		db: memdb.New(),
-	}
-}
-
-func (c *testDB) GetValue(_ context.Context, key []byte) ([]byte, error) {
-	return c.db.Get(key)
-}
-
-func (c *testDB) Insert(_ context.Context, key []byte, value []byte) error {
-	return c.db.Put(key, value)
-}
-
-func (c *testDB) Put(key []byte, value []byte) error {
-	return c.db.Put(key, value)
-}
-
-func (c *testDB) Remove(_ context.Context, key []byte) error {
-	return c.db.Delete(key)
-}
-
-func GetProgramBytes(filePath string) ([]byte, error) {
-	return os.ReadFile(filePath)
-}
-
-func GetGuestFnName(name string) string {
-	return name + "_guest"
-}
